winch: Add full support for integer `sub` and `add` instructions
This patch adds complete support for the `sub` and `add` WebAssembly instructions
for x64, and complete support for the `add` WebAssembly instruction for aarch64.

This patch also refactors how the binary operations get constructed within the
`VisitOperator` trait implementation. The refactor adds methods in the
`CodeGenContext` to abstract all the common steps to emit binary operations,
making this process less repetitive and less brittle (e.g. omitting to push the resulting value
to the stack, or omitting to free registers after used).

This patch also improves test coverage and refactors the filetests directory to make it
easier to add tests for other instructions.

+++ winch/codegen/src/isa/aarch64/asm.rs
@@ -19,8 +19,8 @@ pub(crate) enum Operand {
     Reg(Reg),
     /// Memory address.
     Mem(Address),
-    /// Immediate.
-    Imm(i32),
+    /// 64-bit signed immediate.
+    Imm(i64),
 }
 
 // Conversions between winch-codegen aarch64 types and cranelift-codegen
